chore: move objectUtils to @studio/pure-functions
