+++ frontend/libs/studio-pure-functions/src/types/KeyValuePairs.ts
@@ -1,0 +1,3 @@
+export interface KeyValuePairs<T = any> {
+  [key: string]: T;
+}
